refactor: NetworkTestSetup 생성
